requirements.txt
scikit-learn==1.5.2
python-fasthtml>=0.8.0
matplotlib==3.9.2
scipy==1.14.1
numpy==2.1.2
pandas==2.2.3
pytest
flake8
ipython
# Install the local employee_events package
-e python-package/
